+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ Changelog
 Changes
 -------
 
+* Add derivative metric aggregation for Elasticsearch
 * ...
 
 
@@ -56,6 +57,7 @@ Changes
 * Update release instructions.
 
 Thanks a lot to the contributions from @DWalker487, @bboreham, @butlerx, @dholbach, @franzs, @jaychitalia95, @matthewmrichter and @number492 for this release!
+
 
 0.5.5 (2020-02-17)
 ==================